# docs/api.yaml
openapi: 3.0.0
info:
  version: 1.0-oas3
  title: Barrels of Beer
  contact:
    name: Mateusz Janczak
    url: https://github.com/mateuszjanczak
    email: mateuszjanczak.mj@gmail.com
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - auth-controller
      summary: Logs the user in
      operationId: loginUsingPOST
      requestBody:
        $ref: '#/components/requestBodies/Credentials'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Bad Request
          content:
            application/json:
              examples:
                InvalidRequestBodyError:
                  value:
                    message: Invalid request body.
                    error: BAD_REQUEST
                InvalidPasswordError:
                  value:
                    message: Invalid password.
                    error: BAD_REQUEST
                AccountNotEnabledError:
                  value:
                    message: Account is not enabled.
                    error: BAD_REQUEST
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              examples:
                UserNotFoundError:
                  value:
                    message: User not found.
                    error: NOT_FOUND
  /api/auth/refreshToken/{refreshToken}:
    post:
      tags:
        - auth-controller
      summary: Refreshes the token
      operationId: refreshTokenUsingPOST
      parameters:
        - name: refreshToken
          in: path
          description: refreshToken
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Bad Request
          content:
            application/json:
              examples:
                InvalidArgumentError:
                  value:
                    message: Invalid argument.
                    error: BAD_REQUEST
                InvalidRequestBodyError:
                  value:
                    message: Invalid request body.
                    error: BAD_REQUEST
                TokenNotFoundError:
                  value:
                    message: Token not found.
                    error: BAD_REQUEST
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              examples:
                UserNotFoundError:
                  value:
                    message: User not found.
                    error: NOT_FOUND
  /api/auth/remove/{refreshToken}:
    delete:
      tags:
        - auth-controller
      summary: Removes refresh token
      operationId: removeRefreshTokenUsingPOST
      parameters:
        - name: refreshToken
          in: path
          description: refreshToken
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              examples:
                InvalidArgumentError:
                  value:
                    message: Invalid argument.
                    error: BAD_REQUEST
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /api/taps:
    get:
      tags:
        - tap-controller
      summary: Returns taps
      operationId: getTapsUsingGET
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tap'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - bearerAuth: []
    post:
      tags:
        - tap-controller
      summary: Create tap
      operationId: createTapUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTap'
        description: newTap
        required: true
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
          content:
            application/json:
              examples:
                InvalidRequestBodyError:
                  value:
                    message: Invalid request body.
                    error: BAD_REQUEST
                ValidationError:
                  value:
                    message: Validation failed.
                    error: BAD_REQUEST
                    validation:
                      tapId:
                        - Tap id must be a positive number.
                TapAlreadyExistsError:
                  value:
                    message: Tap already exists.
                    error: BAD_REQUEST
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - bearerAuth: []
  /api/taps/{tapId}:
    get:
      tags:
        - tap-controller
      summary: Returns the selected tap
      operationId: getTapByIdUsingGET
      parameters:
        - name: tapId
          in: path
          description: tapId
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tap'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              examples:
                TapNotFoundError:
                  value:
                    message: Tap not found.
                    error: NOT_FOUND
      security:
        - bearerAuth: []
    post:
      tags:
        - tap-controller
      summary: Sets the selected tap
      operationId: setTapUsingPOST
      parameters:
        - name: tapId
          in: path
          description: tapId
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TapDetails'
        description: tapDetails
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              examples:
                InvalidArgumentError:
                  value:
                    message: Invalid argument.
                    error: BAD_REQUEST
                InvalidRequestBodyError:
                  value:
                    message: Invalid request body.
                    error: BAD_REQUEST
                ValidationError:
                  value:
                    message: Validation failed.
                    error: BAD_REQUEST
                    validation:
                      capacity:
                        - Capacity must be a positive number.
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              examples:
                TapNotFoundError:
                  value:
                    message: Tap not found.
                    error: NOT_FOUND
      security:
        - bearerAuth: []
...
servers:
  - url: http://localhost:8080/
components:
  requestBodies:
    Credentials:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Credentials'
      description: credentials
      required: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Credentials:
      type: object
      required:
        - password
        - username
      properties:
        password:
          type: string
        username:
          type: string
      title: Credentials
    NewTap:
      type: object
      properties:
        tapId:
          type: integer
          format: int32
      title: NewTap
...
    Tap:
      type: object
      required:
        - barrelContent
        - capacity
        - currentLevel
        - enabled
        - tapId
        - temperature
      properties:
        barrelContent:
          type: string
        capacity:
          type: integer
          format: int64
        currentLevel:
          type: integer
          format: int64
        enabled:
          type: boolean
        tapId:
          type: integer
          format: int32
        temperature:
          type: number
          format: float
      title: Tap
    TapDetails:
      type: object
      required:
        - contentType
      properties:
        capacity:
          type: integer
          format: int64
        contentType:
          type: string
          enum:
            - CHMYZ_Pils
            - GAZDA_Marcowe
            - KADUK_Podwójny_Koźlak
            - KICARZ_Koźlak
            - KRASA_Weizen
            - SĘDEK_IPA
            - UPIR_Dunkel
      title: TapDetails
...
    Token:
      type: object
      required:
        - expirationTime
        - refreshToken
        - token
      properties:
        expirationTime:
          type: integer
          format: int64
        refreshToken:
          type: string
        token:
          type: string
      title: Token
...